+++ src/Writer.php
@@ -14,7 +14,8 @@ class Writer
         private Formatter $formatter,
         private Config $config,
         private FileName $fileName
-    ) {}
+    ) {
+    }
 
     /**
      * Write a query to log.
@@ -73,7 +74,7 @@ class Writer
         return $this->config->queriesEnabled() &&
             $query->time() >= $this->config->queriesMinExecTime() &&
             preg_match($this->config->queriesIncludePattern(), $query->raw()) &&
-            !preg_match($this->config->queriesExcludePattern(), $query->raw());
+            ! preg_match($this->config->queriesExcludePattern(), $query->raw());
     }
 
     /**