+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [1.0.x (Unreleased)](https://github.com/onlime/laravel-sql-reporter/compare/v1.0.0...main)
 
+- Allow bindings to be null.
+
 ## [v1.0.0 (2022-02-10)](https://github.com/onlime/laravel-sql-reporter/releases/tag/compare/v0.9.1...v1.0.0)
 
 - Support Laravel 9